utils rename to graph_utils (name conflict)
